# phase-3/specs/001-phase3-todo-chatbot/contracts/openapi.yaml
openapi: 3.0.3
info:
  title: Todo AI Chatbot API
  description: API for the Todo AI Chatbot application with natural language processing
  version: 1.0.0
servers:
  - url: http://localhost:8000/api
    description: Local development server
  - url: https://api.todo-chatbot.com
    description: Production server
paths:
  /{user_id}/chat:
    post:
      summary: Send a message to the AI chatbot
      description: Process user message through AI agent and return response
      parameters:
        - name: user_id
          in: path
          required: true
          description: The ID of the authenticated user
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                conversation_id:
                  type: integer
                  description: Existing conversation ID (creates new if not provided)
                  example: 123
                message:
                  type: string
                  description: User's natural language message
                  example: "Add a task to buy groceries"
              required:
                - message
      responses:
        '200':
          description: Successful response from AI assistant
          content:
            application/json:
              schema:
                type: object
                properties:
                  conversation_id:
                    type: integer
                    description: The conversation ID
                    example: 123
                  response:
                    type: string
                    description: AI assistant's response
                    example: "I've added 'Buy groceries' to your task list"
                  tool_calls:
                    type: array
                    description: List of MCP tools invoked
                    items:
                      type: object
                      properties:
                        name:
                          type: string
                          description: Name of the tool called
                          example: "add_task"
                        parameters:
                          type: object
                          description: Parameters passed to the tool
                          example: {"user_id": "user123", "title": "Buy groceries"}
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '422':
          $ref: '#/components/responses/ValidationError'
...
components:
...
  responses:
    Unauthorized:
      description: Missing or invalid authentication token
      content:
        application/json:
          schema:
            type: object
            properties:
              detail:
                type: string
                example: "Not authenticated"
    Forbidden:
      description: Access forbidden - attempting to access another user's data
      content:
        application/json:
          schema:
            type: object
            properties:
              detail:
                type: string
                example: "Access forbidden"
...
    ValidationError:
      description: Request validation failed
      content:
        application/json:
          schema:
            type: object
            properties:
              detail:
                type: array
                items:
                  type: object
                  properties:
                    loc:
                      type: array
                      items:
                        type: string
                    msg:
                      type: string
                    type:
                      type: string
                example: [{"loc": ["body", "title"], "msg": "Field required", "type": "value_error"}]
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: JWT token from Better Auth system
security:
  - BearerAuth: []